requirements.txt
autopep8==1.6.0
mysql-connector-python==8.0.27
protobuf==3.19.1
pycodestyle==2.8.0
python-dotenv==0.19.2
toml==0.10.2
